Remove undefined behavior from bit-shifting code
PR #536 

Overflow of signed integer arithmetic (including shifting) is
undefined behavior in C++. Compilers are allowed to exploit
this UB for optimization purposes up to and including not
emitting any opcodes.

These were found using cppcheck.

+++ instructionAPI/src/InstructionDecoder-aarch64.h
@@ -207,7 +207,7 @@ namespace Dyninst {
 
             int highest_set_bit(int32_t val) {
                 for (int bit_index = 31; bit_index >= 0; bit_index--)
-                    if (((val >> bit_index) & 0x1) == 0x1)
+                    if (((static_cast<uint32_t>(val) >> bit_index) & 0x1) == 0x1)
                         return bit_index + 1;
 
                 return -1;
@@ -215,7 +215,7 @@ namespace Dyninst {
 
             int lowest_set_bit(int32_t val) {
                 for (int bit_index = 0; bit_index <= 31; bit_index++)
-                    if (((val >> bit_index) & 0x1) == 0x1)
+                    if (((static_cast<uint32_t>(val) >> bit_index) & 0x1) == 0x1)
                         return bit_index + 1;
 
                 return -1;
